:wrench: Remake of GVars
The GVar (global variable) system needed some improvements:
* Use char buffers instead of std::string - plays nicely with DearIMGUI's text input boxes.
* Generalize the active/pending value mechanism to all GVars - it will become necessary in the future when GVars will be read on sim-threads and updated on mainthread.
* Create GVar object and array of these - will be useful for {printing all / searching through} GVars in the console.

The "GameSettings" GUI window was removed - will be replaced by leaner IMGUI-based one.
Also, "MultiplayerSelector" GUI was replaced by a dummy - new IMGUI code is already prepared in my testbed project.

+++ source/main/gui/panels/GUI_GameMainMenu.h
@@ -43,8 +43,8 @@ public:
     GameMainMenu();
 
     // Keyboard updates - move up/down and wrap on top/bottom. Initial index is '-1' which means "no focus"
-    inline void   KeyUpPressed()                   { m_kb_focus_index = (m_kb_focus_index <= 0) ? (NUM_BUTTONS-1) : (m_kb_focus_index - 1); }
-    inline void   KeyDownPressed()                 { m_kb_focus_index = (m_kb_focus_index < (NUM_BUTTONS - 1)) ? (m_kb_focus_index + 1) : 0; }
+    inline void   KeyUpPressed()                   { m_kb_focus_index = (m_kb_focus_index <= 0) ? static_cast<int>(NUM_BUTTONS-1) : (m_kb_focus_index - 1); }
+    inline void   KeyDownPressed()                 { m_kb_focus_index = (m_kb_focus_index < static_cast<int>(NUM_BUTTONS - 1)) ? (m_kb_focus_index + 1) : 0; }
     inline bool   IsVisible() const                { return m_is_visible; }
     inline void   SetVisible(bool v)               { m_is_visible = v; }
     inline bool   CheckKbEnter(int idx)            { return (idx == m_kb_focus_index) && (ImGui::GetIO().KeysDown[ImGuiKey_Enter]); }
